requirements.txt
# LLM 및 에이전트 처리용
openai>=1.3.5
langgraph>=0.4.0

# 보고서 생성 (HTML → PDF)
weasyprint>=60.0

# 마크다운 변환
markdown2>=2.4.12

# 시각화 (재무 차트 등)
matplotlib>=3.7.1

# 웹 UI
streamlit>=1.30.0

langchain_community
langchain_openai